async_rollout_worker_passed
